refactor: move hls logic to its own module
